Rename PDF => PDFIntegration and move to `src/annotator/integrations`
Create an `integrations` directory within `src/annotator` which will
house the `*Integration` classes that encapsulate document
type/viewer-specific functionality, and create the first version of the
PDF integration by moving/renaming the existing `PDF` class. The
supporting `PDFMetadata` class has also been moved.

Per recently agreed code conventions [1], the `PDFIntegration` and
`PDFMetadata` classes are no longer default exports.

[1] https://github.com/hypothesis/frontend-toolkit/blob/master/docs/js-guide.md#default-exports
